Put license in name of Kim Holder
Since Moonwards isn't a formal entity, there is no point in putting it in the name of Moonwards, i don't think. I suppose if later it can be defined who Moonwards is, then we can change the license to say Moonwards, and that would mean the members of Moonwards according to some agreed upon list.

Former-commit-id: bf8d2b0a846752923a78d2bf2074710a245c5201