Styling with columns
Layout on Moonwards pages uses a column format designed to make it responsive. Divs and headings not given column classes will take the width of their contents and not wrap. Because of the way the CSS is written, it is necessary to stick to some layout conventions, such as placing headings in their own rowR or rowL div.

Former-commit-id: 673def831c7f2995df3bbeb1b8ed56b1c1efcf4d